feat:add handler for add buttons
