+++ online store/src/components/controller/AppController.ts
@@ -12,7 +12,6 @@ class AppController {
         // отрисовывает данные при загрузке страницы
         // вызывает хендлер при добавлении продукта в корзину
         // this.resetData(this.model.data, this.model.cart.length)
-        this.view.bindAddProduct(this.handleAddProduct);
         // this.view.search.bindSearchProduct(this.handlerSearchProduct);
         // this.view.categories.bindAddCategory(this.handlerAddCategory);
     }
@@ -23,6 +22,8 @@ class AppController {
         this.view.price.drawPrice(this.model.getTotalSum().toString());
         this.view.categories.drawCategories(this.model.categories);
         this.view.search.drawSearch();
+        
+        this.view.card.bindAddProduct(this.handleAddProduct);
         // this.view.displayContent(this.model.data);
         // this.view.categories.drawCategories(this.model.categories);
         // this.view.card.drawCard(this.model.data);
@@ -39,8 +40,6 @@ class AppController {
                 urlRoute(e, anchor.id);
             })
         })
-        console.log('HOME PAGE')
-
     }
     //  сама функция отрисовки  категорий
     // resetCategories(arr: CategoriesProduct[]) {
@@ -70,11 +69,9 @@ class AppController {
     // };
     //  функция вызывается при добавлении продукта и закидывает продукт в массив корзины.Перерисовка страницы с новыми данными
     handleAddProduct = (id: number) => {
-        console.log('Handler');
         this.model.addProduct(id);
-        this.view.price.drawPrice(this.model.getTotalSum().toString());
         this.startPage();
-
+        
     };
     // для сортировки товара по тексту введенном в инпуте
     // handlerSearchProduct = (textInput: string) => {
