+++ online store/src/components/view/CartView.ts
@@ -1,28 +1,18 @@
 import { Product } from "../../types/types";
 import Cart from "./cart/cart";
 import PageView from "./Pages/MainComponent/PageView";
+import ProductsFromCart from "./СartView/ProductsFromCart/ProductsFromCart";
 
 
 class CartView {
   mainPage: PageView;
   cart: Cart;
+  products: ProductsFromCart;
   constructor() {
     this.mainPage = new PageView();
     this.cart = new Cart();
+    this.products = new ProductsFromCart();
   }
-  drawCart(data: Product[]) {
 
-    const container = document.querySelector('main .container');
-    while (container?.firstChild) {
-      container.removeChild(container.firstChild)
-    }
-    const cartBlock = document.createElement('div');
-    const cartContent =
-      `
-    <p>${data.length}</p>
-    `
-    cartBlock.innerHTML = cartContent;
-    container?.append(cartBlock);
-  }
 }
 export default CartView;
