+++ online store/src/components/view/СartView/ProductsFromCart/ProductsFromCart.module.sass
@@ -1,0 +1,3 @@
+.cartBlock
+  width: 400px
+  height: 200px
