+++ online store/src/components/view/СartView/ProductsFromCart/ProductsFromCart.ts
@@ -1,0 +1,28 @@
+import { Product } from './../../../../types/types';
+import classes from './ProductsFromCart.module.sass';
+
+
+class ProductsFromCart {
+  drawCarts(data: Product[]) {
+    const container = document.querySelector('main .container');
+    while (container?.firstChild) {
+      container.removeChild(container.firstChild)
+    }
+    data.forEach(el => {
+      const cartBlock = document.createElement('div');
+      cartBlock.className = `${classes.cartBlock}`
+      const cartContent =
+        `
+      <h3>${el.title}</h3>g
+      <img src='${el.thumbnail}'
+      <p>${el.description}</p>
+      `
+      cartBlock.innerHTML = cartContent;
+      container?.append(cartBlock);
+    })
+
+  }
+}
+
+
+export default ProductsFromCart;
